naming adjustments

+++ DragonFruit.Orbit.API/Objects/User/OsuUserScoreInfoStatistics.cs
@@ -8,21 +8,21 @@ namespace DragonFruit.Orbit.API.Objects.User
     public class OsuUserScoreInfoStatistics
     {
         [JsonProperty("count_50")]
-        public long Count50 { get; set; }
+        public uint Count50 { get; set; }
 
         [JsonProperty("count_100")]
-        public long Count100 { get; set; }
+        public uint Count100 { get; set; }
 
         [JsonProperty("count_300")]
-        public long Count300 { get; set; }
+        public uint Count300 { get; set; }
 
         [JsonProperty("count_geki")]
-        public long CountGeki { get; set; }
+        public uint CountGeki { get; set; }
 
         [JsonProperty("count_katu")]
-        public long CountKatu { get; set; }
+        public uint CountKatu { get; set; }
 
         [JsonProperty("count_miss")]
-        public long CountMiss { get; set; }
+        public uint CountMiss { get; set; }
     }
 }
